Add files via upload
Here is the set of insights I have come across after going through the dataset:
The highest no. of movies (3052) were produced in the year 2017. And if we consider the monthly trends; the maximum
number of movies released in March (824) and the lowest in December (438) in a span of three years 2017 to 2019. And out
of all 13 genres, the maximum number of movies were from the Drama genre with an average of 106 minutes duration.
RSVP should consider the following factors while producing its next movie:
 The Genre should be preferred as Drama and the duration should be around 106 minutes.
 The Director should be James Mangold, as he has the highest average rating of 8+ among the top 3
genres with max movie count.
 The Production House, RSVP can partner with Dream Warrior Pictures or National Theatre Live for the next
project, as these two production companies have the highest hit movie counts with an average rating of 8.
 For The Lead Roles If RSVP has a script for:
The worldwide audience then they should consider Mammootty. Has a worldwide fanbase and has 8 movies with
a median rating > 8.
Indian audience then should cast Vijay Sethupathi and Taapsee Pannu. Their weighted average rating is
highest with 8.42 and 7.70.
 The Timing and RSVP must avoid March and prefer December to get maximum coverage since competing
movies are very few.
 If they want to go for a global partnership; Marvel Studios (Rank 1), Twentieth Century Fox (2) or
Warner Bros.(3) can be chosen. They have garnered the highest number of votes.